File writing and saving
Added the ability to create "Code.txt" on launch if it has not already
been created. Also, the ability to overwrite the text file and save the
text file. Finally, similar to TextEdit or Sublime, the file with save
when entering the background or when the app is terminating. Therefore,
you will never lose progress on your code. Future idea, add an auto
save feature that saves every 30 seconds for random crashes.